Simplify theme and improve performance
Simplify things a little bit so that things take as little space as
possible: encode return code and git branch status with colors, and
remove extra characters that are there only for aesthetics.

Also, limit the scope of `git status` so that it only looks for changed
files. While some information will be missing (it won't detect new
files) it is noticeably faster on huge repositories (e.g. Liferay).